Allow independent credentials for JSON RPC
It's probably not a great idea to allow users to easily access JSON RPC,
so a different set of credentials is useful.

Also allow the secret to be mounted directly without the convoluted
auth-config format (which is an ini file in reality). Deprecate
the auth-config approach (JSON RPC is the last instance where it's
used).

Signed-off-by: Dmitry Tantsur <dtantsur@protonmail.com>